requirements.txt
openai-agents>=0.0.7
openai>=1.0.0
streamlit>=1.32.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
